changes and additions

+++ api/src/posts/posts.controller.ts
@@ -14,7 +14,7 @@ import { Queue } from 'bullmq';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { ChatGateway } from 'src/chat/chat.gateway';
 import { ZodValidationPipe } from 'src/zod-validation.pipe';
-import { BookmarkPost, BookmarkPostDTO, BulkViewPost, BulkViewPostDTO, CreatePost, CreatePostDTO, CreateReel, CreateReelDTO, CreateSharedPost, CreateSharedPostDTO, DeletePost, DeletePostDTO, GetBookmarkedPostsDTO, GetPost, GetPostDTO, GetPostLikes, GetPostLikestDTO, GetPromotions, GetPromotionsDTO, LikeCommentOrReply, LikeCommentOrReplyDTO, LikePost, LikePostDTO, PromotePost, PromotePostDTO, PromotionActivation, PromotionActivationDTO, ReportPost, ReportPostDTO, UpdatePost, UpdatePostDTO, ViewPost, ViewPostDTO } from 'src/schema/validation/post';
+import { BookmarkPost, BookmarkPostDTO, BulkViewPost, BulkViewPostDTO, CreatePost, CreatePostDTO, CreateSharedPost, CreateSharedPostDTO, DeletePost, DeletePostDTO, GetBookmarkedPostsDTO, GetPost, GetPostDTO, GetPostLikes, GetPostLikestDTO, GetPromotions, GetPromotionsDTO, LikeCommentOrReply, LikeCommentOrReplyDTO, LikePost, LikePostDTO, PromotePost, PromotePostDTO, PromotionActivation, PromotionActivationDTO, ReportPost, ReportPostDTO, UpdatePost, UpdatePostDTO, ViewPost, ViewPostDTO } from 'src/schema/validation/post';
 import { Request } from 'types/global';
 import { Cursor, CursorDTO, ValidMongoId } from 'src/schema/validation/global';
 import Stripe from 'stripe';
@@ -101,9 +101,9 @@ export class PostsController {
     @Get('reels')
     async getReels(@Req() req: Request, @Res() response: Response) {
         const { sub } = req.user
-        const { cursor, targetId } = req.query as { type: string, cursor: string, targetId: string }
+        const { cursor, targetId, type } = req.query as { type: string, cursor: string, targetId: string }
         console.log(targetId, 'get reels targetId')
-        response.json(await this.postService.getReels(cursor, sub, targetId, 'user'))
+        response.json(await this.postService.getReels(cursor, sub, targetId, type))
     }
 
     // @Public()
@@ -264,40 +264,57 @@ export class PostsController {
             throw new BadRequestException("You are not allowed to delete this post")
         }
 
-        post.media.forEach(async (media) => {
-            if (media.url) {
-                let videoUrlSplit = media.url.split("/")
-                let thumbnailUrlSplit = media.thumbnail.split("/")
-                let watermarkUrlSplit = media.watermarkUrl.split("/")
+        try {
 
-                console.log(`deleting ${videoUrlSplit}, ${thumbnailUrlSplit}, ${watermarkUrlSplit} from s3...`)
 
-                let videoFilename = videoUrlSplit[videoUrlSplit.length - 1]
-                let thumbnailFilename = thumbnailUrlSplit[thumbnailUrlSplit.length - 1]
-                let watermarkedVideoFilename = watermarkUrlSplit[watermarkUrlSplit.length - 1]
+            post.media.forEach(async (media) => {
+                if (media.url) {
+                    let videoUrlSplit = media?.url?.split("/")
+                    let thumbnailUrlSplit = media?.thumbnail?.split("/")
+                    let watermarkUrlSplit = media?.watermarkUrl?.split("/")
 
-                console.log(videoFilename, thumbnailFilename, watermarkedVideoFilename, 'filenames')
+                    console.log(`deleting ${videoUrlSplit}, ${thumbnailUrlSplit}, ${watermarkUrlSplit} from s3...`)
 
-                console.log(`deleting ${videoFilename} from s3...`)
-                const video = await this.uploadService.deleteFromS3(videoFilename)
+                    let videoFilename = videoUrlSplit?.[videoUrlSplit?.length - 1]
+                    let thumbnailFilename = thumbnailUrlSplit?.[thumbnailUrlSplit?.length - 1]
+                    let watermarkedVideoFilename = watermarkUrlSplit?.[watermarkUrlSplit?.length - 1]
 
-                console.log(`deleting ${thumbnailFilename} from s3...`)
-                const thumbnail = await this.uploadService.deleteFromS3(thumbnailFilename)
+                    console.log(videoFilename, thumbnailFilename, watermarkedVideoFilename, 'filenames')
 
-                const watermarkVideo = await this.uploadService.deleteFromS3(watermarkedVideoFilename)
-                console.log(`deleting ${watermarkedVideoFilename} from s3...`)
-                console.log(video.$metadata.httpStatusCode, thumbnail.$metadata.httpStatusCode, watermarkVideo.$metadata.httpStatusCode, 'deleted')
-            }
-        })
+                    console.log(`deleting ${videoFilename} from s3...`)
 
-        const deleted = await this.postService.deletePost(deleteReelData.postId)
-        res.json({ deleted: true, success: true })
+                    if (videoFilename) {
+                        const video = await this.uploadService.deleteFromS3(videoFilename)
+                        console.log(video.$metadata.httpStatusCode)
+                    }
+
+                    if (thumbnailFilename) {
+                        const thumbnail = await this.uploadService.deleteFromS3(thumbnailFilename)
+                        console.log(`deleting ${thumbnailFilename} from s3...`)
+                        console.log(thumbnail.$metadata.httpStatusCode)
+                    }
+
+                    if (watermarkedVideoFilename) {
+                        const watermarkVideo = await this.uploadService.deleteFromS3(watermarkedVideoFilename)
+                        console.log(`deleting ${watermarkedVideoFilename} from s3...`)
+                        console.log(watermarkVideo.$metadata.httpStatusCode, 'deleted')
+                    }
+
+
+                }
+            })
+
+            const deleted = await this.postService.deletePost(deleteReelData.postId)
+            res.json({ deleted: true, success: true })
+        } catch (error) {
+            throw new BadRequestException("Error deleting reel post, please try again later.");
+        }
     }
 
     @UseInterceptors(FileInterceptor('file'))
     @Post("create/reel")
     async createReel(
-        @Body(new ZodValidationPipe(CreateReel, true, "reelData")) reelData: CreateReelDTO,
+        @Body(new ZodValidationPipe(CreatePost, true, "reelData")) reelData: CreatePostDTO,
         @Req() req: Request,
         @Res() res: Response,
         @UploadedFile() file: Express.Multer.File
@@ -310,8 +327,8 @@ export class PostsController {
         const filename = uuidv4();
         const uploadPromise = [this.uploadService.processAndUploadContent(file.buffer, filename, fileType, file.originalname, true)];
 
-        const { sub } = req.user;
-        let targetId = new Types.ObjectId(sub);
+        const { sub } = req.user
+        let targetId = reelData.type == "user" ? new Types.ObjectId(sub) : new Types.ObjectId(reelData.targetId)
 
         let uploadedPost = await this.postService.createPost({
             ...reelData,
